requirements.txt
Avalara
cryptography==3.3.2
pyOpenSSL==20.0.0
stripe
astor==0.8.1
odoo_upgrade @ git+https://github.com/odoo/upgrade-util@master